+++ tests/unit/models/FolderTest.php
@@ -3,13 +3,13 @@
 namespace davidhirtz\yii2\media\tests\unit\models;
 
 use Codeception\Test\Unit;
-use davidhirtz\yii2\media\models\Folder;
+use davidhirtz\yii2\media\models\collections\FolderCollection;
 
 class FolderTest extends Unit
 {
     public function testFolder(): void
     {
-        $folder = Folder::getDefault();
+        $folder = FolderCollection::getDefault();
         $this->assertFalse($folder->getIsNewRecord());
     }
 }